% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/male_config.R
\docType{data}
\name{male_config}
\alias{male_config}
\title{Config file for spatial prioritization in Male, Maldives}
\format{
Simple feature
}
\source{
The data frame compiles data from OSM and UN-OCHA.
}
\usage{
data(male_config)
}
\description{
Config file for spatial prioritization in Male, Maldives
}
\examples{
data(male_config)
}
\references{
Data from CAPSUS, 2022, Climate-Smart Scenarios for
Affordable Housing in Maldives.
}
\keyword{configuration}
\keyword{data}
\keyword{sets}
