% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/male_spatial.R
\docType{data}
\name{male_spatial}
\alias{male_spatial}
\title{Isochrones and isodistances from Male, Maldives}
\format{
Simple feature
}
\source{
The data frame compiles data from OSM and UN-OCHA.
}
\usage{
data(male_spatial)
}
\description{
Isochrones and isodistances from Male, Maldives
}
\examples{
data(male_spatial)
}
\references{
Data from CAPSUS, 2022, Climate-Smart Scenarios for
Affordable Housing in Maldives.
}
\keyword{configuration}
\keyword{data}
\keyword{sets}
