\docType{package}
\name{prioritization-package}
\alias{prioritization-package}
\title{Routines for spatial prioritization for alien clearing}
\description{
  This package is a collection of different spatial
  prioritization routines to prioritize areas spatially
}
\details{
  \tabular{ll}{ Package: \tab prioritization\cr Type: \tab
  Package\cr Version: \tab
  MAJORVERSION.MINORVERSION.PATCHVERSION\cr Date: \tab
  TODAYSDATE\cr License: \tab GPL (>= 2)\cr LazyLoad: \tab
  yes\cr }
}
\author{
  Rainer M Krug \email{Rainer@krugs.de}
}

