% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prioritization.R
\name{prioritization}
\alias{prioritization}
\title{Prioritization index}
\usage{
prioritization(x, config)
}
\arguments{
\item{x}{data frame with numeric or integer variables.}

\item{config}{data frame with standardization rules.}
}
\value{
Data frame with standardized values and a prioritization index.
}
\description{
Creates an index from each indicator (variable) in a data frame.
}
\examples{
# Example of how can be used this function.
data(male_spatial)
data(male_config)
prioritization(x = male_spatial, config = male_config)

}
