% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_pupop.R
\name{s_popup}
\alias{s_popup}
\title{Pop ups}
\usage{
s_popup(x, r, my_variables, pretty_names, digits = 0)
}
\arguments{
\item{x}{parameter name and vector type.}

\item{r}{parameter name and vector type.}

\item{my_variables}{parameter name and vector type.}

\item{pretty_names}{parameter name and vector type.}

\item{digits}{(default 0).}
}
\value{
Vector kind output by the function and short description if is necessary.
}
\description{
Create a character vector that summarizes and formats all variables in a data frame.
}
\examples{
# Example of how can be used this function.
# Code lines

}
