% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_standardize.R
\name{s_standardize}
\alias{s_standardize}
\title{Short title of the function given what it does}
\usage{
s_standardize(x, selected_method = "observe", MIN = NULL, MAX = NULL)
}
\arguments{
\item{x}{data frame of spatial polygons.}

\item{selected_method}{character vector.}

\item{MIN}{(default NULL).}

\item{MAX}{(default NULL).}
}
\value{
Vector kind output by the function and short description if is necessary.
}
\description{
Short description of what the function does.
}
\examples{
# "benchmark" as selected method.
s_standardize(
x = c(0, 50, 100, 200),
selected_method = "benchmark",
MIN = 0,
MAX = 100)

# "observe" as selected method
s_standardize(
x = c(-100, 50, 100, 200),
selected_method = "observe",
MIN = 0,
MAX = 100)

}
