\docType{package}
\name{prioritization-package}
\alias{prioritization-package}
\title{Routines for spatial prioritization for alien clearing}
\description{
  This package is a collection of different spatial
  prioritization routines to prioritize areas spatially
}
\details{
  \tabular{ll}{ Package: \tab prioritization\cr Type: \tab
  Package\cr Version: \tab 0.1.0\cr Date: \tab
  2013-10-01_10-01\cr License: \tab GPL (>= 2)\cr LazyLoad:
  \tab yes\cr }
}
\author{
  Rainer M Krug \email{Rainer@krugs.de}
}

