% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_config.R
\name{s_config}
\alias{s_config}
\title{s_config}
\usage{
s_config(
  x,
  name = NA,
  group = NA,
  weight = NA,
  direction = NA,
  method = NA,
  min = NA,
  max = NA
)
}
\arguments{
\item{x}{data frame.}

\item{name}{human friendly names for each variable in x. Defaults to column names in x.}

\item{group}{vector with the grouping variable  with the same length as rows in x. Defaults to a single "general" group.}

\item{weight}{vector with the weights that that will be passed to weighted mean. The vector might have a single character, or a vector with the same length as rows in x. Defaults to 1.}

\item{direction}{vector stating if each variable in x "larger is better" or "smaller is better". The vector might have a single character, or a character vector with the same length as rows in x. Defaults to "smaller is better".}

\item{method}{vector stating the standardization method should be observe or benchmark. The vector might have a single character or a character vector with the same length as rows in x. Defaults to "observe".}

\item{min}{vector with the minimum value that should be considered in the benchmark method. The vector might have a single number or a numeric vector with the same length as rows in x. Defaults to the min value for each variable in x.}

\item{max}{vector with the maximum value that should be considered in the benchmark method. The vector might have a single number or a numeric vector with the same length as rows in x. Defaults to the max value for each variable in x.}
}
\value{
A config data frame.
}
\description{
Creates an config data frame from a data frame or sf.
}
\examples{
data(male)
x <- sf::st_drop_geometry(male_spatial)
s_config(x)

}
